Add a task decorator to handle actions execution and revert
The 'task' class is a Generic[A], where 'A' is the prototype for the
action function to be turned as a task. The type system ensures that the
decorated @task.revert function matches that prototype, as expected by
the rollback logic.

This adds tests, so we setup the infrastructure accordingly.

In tests/, we configure mypy to allow untyped defs. This is a good
compromise to allow type checking in tests while not requiring all test
functions to be annotated.