Add an 'instance' module with an 'init()' task
We implement calls to initdb, with a revert mechanism using our task
implementation.

Actual call to initdb executable is simply done by
subprocess.check_call() at the moment (and until we actually need
something else). This mean that standard output will currently be
redirected to the parent process (that of python).

Also notice that we use 'pg_config' to determine the bindir of
postgresql.

In CI configuration, we now install postgresql and add an unprivileged
user in order to run initdb.