Add instance configuration task
We add a configure() task to write a custom postgresql configuration
file, with a chosen name, that is included on top of default
postgresql.conf.

At the moment, configuration items are left free to be defined by the
caller and we only set 'cluster_name' based on instance name.

This adds a dependency on pgtoolkit.