Make instance.init() idempotent
We introduce Instance.exists() to do a system lookup for an existing
instance, checking consistency along the way, and possible return early
in instance.init() if the instance already exists.

In test_init(), we now use version=11 in order to match what will be
installed on Debian/buster (as is in CI).