Build and user our own CI image
We add a "ci_image" job, running at "build" stage that builds a
Docker image suitable for our tests and pushes it to the local container
registry.

This image installs postgresql and pgbackrest from pgdg repository. We
need that repo to get pgbackrest in a version compatible with our
expectation. The image also configures a non-root user so that our tests
would not not need to "su" anymore.

By defining stages "build" and "test" in .gitlab-ci.yml, we get
automatic dependency and can use the CI image from "build" stage in jobs
from the "test" stage.