Introduce execution "context" protocols and implementations
We introduce a base context class and a concrete implementation in 'ctx'
module. This class holds a run() method, previously defined as
run_command(). For the Ansible interface, an AnsibleContext is also
provided.

This context object is passed to all tasks of 'instance' and
'pgbackrest' modules (even those that do not make use of it at the
moment, for consistency).

The default Context uses a run() function defined in cmd module.
Previously it was an alias to pgtoolkit.ctl.run_command, be now the code
is copied over to keep more control over it.