Define hooks to setup components on the instance
We add the infrastructure to use pluggy as a plugin manager and use it
to declare hooks on instance events (such as creation, deletion,
(re-)configuration). These hooks are used to call (pgbackrest and
prometheus) components' tasks.

The plugin manager is declared in a "pm" module as a simple wrapper to
pluggy default class with a couple of helper methods. The plugin manager
is attached to the context and then used in instance tasks to call hooks
where appropriate (i.e. after instance configuration, before instance
drop, after instance startup).

Hook specifications are declared in hookspecs module. Hook
implementations are in the pgbackrest and prometheus module.

In Ansible module, it is no longer needed to call pgbackrest tasks now.

In test_instance.py, we unregister plugins to avoid side-effects and
only test instance setup with side components. Similarly, in
pgbackrest's tests (resp. prometheus's tests) we unregister respective
hooks since we call tasks explicitly.

Note that, since pluggy is not typed, we need a couple of '# type:
ignore' here and there.