Move task revert logic into a runner() context manager
(This may not be bullet-proof, but it's simple and seems to be working.)

The new 'runner' context manager maintain a stack of task calls and then
revert them in reverse call order.

In test_instance.py, we were previously relying on task-revert logic, so
we now use a runner explicitly.

The only places where a runner appears needed is in instance.apply() and
the Ansible module.