Extend JSON loading of settings to support serialized values
We change the meaning of SETTINGS environment variable to support
'@<file.json>' when custom settings are meant to be retrieved from a
file and consider other values (not starting with '@') to be serialized
JSON.

We'll need this in tests in the next commits in order to have systemd
working with the 'postgres' wrapper script introduced earlier.

Along the way, we add tests.