Make postgresql version configurable
The new 'versions' field in settings.postgresql is now a mapping from
version string to PostgreSQLVersionSettings instance, the latter being a
new (minimal) model. The 'default_version' allows one to define which
version is the default, if unspecified at runtime. Loading and default
values definition is a bit tricky and is handled with the set_versions()
root validator.