Handle password for postgresql surole
We turn the "surole" string field into a dedicated SuRole object, which
includes the role name and optional role password. By setting the
env_prefix, we can set the password from the environment. We also
disable the "secrets" file source, as it'd be quite unwieldy to use.

InitdbSettings is expanded to allow configuration of local and host
authentication rules. These settings are respectively used as options
for initdb call.

If a super-user password is supplied, the configure_auth() task will
update the role to set its password. For this, we need to temporarily
disable auth-local hba rule in order to allow connection. The operation
is not idempotent, see TODO inline. The configure_auth() task is called
in apply() task.

Password encryption method is now configurable through
postgresql.password_encryption setting, which is used to set respective
setting in postgresql configuration file, during configure() task.