Add a queries module to handle SQL queries
This module exposes a get() function which will look for a named query
in queries.sql file located next to the module. The iter() function will
load all queries. We start with the query to set the password of a role,
that will be needed shortly. The idea is similar to anosql, but with
a more limited feature-set.

In tests, we add a --regen-test-data option to automatically update test
data when the queries.sql file will change.