Fix connection settings for prometheus
The previous attempt was wrong because DATA_SOURCE_PASS_FILE is not a
pgpass file, but a plain file containing only the password. While
investigating, it appears that we need to disable sslmode explicitly
because go's libpq apparently sets this by default. Finally, we bundle
everything into DATA_SOURCE_NAME environment variable, written in
postgres_exporter configuration file.

To make sure this works, we complement functional tests to check that
the web service is up and responding with metrics.