Split Instance model into InstanceSpec(BaseInstance)
The new InstanceSpec is meant to hold an instance to be created, in
contrast with Instance that should concern an existing instance.
Accordingly, the config() method of Instance now always returns a
Configuration object and we can drop a lot of 'assert config'.

In instance module, we now use either classes when appropriate (e.g.
InstanceSpec is used as input of init()). Some functions now return a
transformed Instance object (built from InstanceSpec). Some functions,
like describe() or drop(), can operate on a non-existent instance and
return None in that case. We factor out this behavior in a
bypass_absent_instance() decorator.

Conversion from and to InstanceSpec is performed by methods of the
Instance class.

Code of BaseInstance.exists() is slightly different from before as it
now raises a LookupError in case of version mismatch.