Handle pgpass update upon instance port change
We refactor roles.instance_configure() hook so that it handles all roles
matching "old" port value in the password file by updating their entry
upon port change. The super-user role is still handled specifically
(entry creation) in this hook, so we make this clearer in the docstring.

In tests, we extract role_in_pgpass() helper function and make it more
generic so that it can be reused in newly added test case.