Add a 'port' field to Instance's manifest and Ansible param spec
The field is optional so that we do not write any specific configuration
if unspecified. This value is passed to instance.configure() from
instance.apply(). The None value is handled there already, per previous
commit. In instance.describe(), we fill the (output) manifest with the
real instance port and take care of dropping the 'port' field from
managed configuration dict.

In the Ansible module, the parameters spec is updated to add a 'port'
field; the same logic as in manifest/instance.apply() is then implemented.