Allow site-wise settings definition through a YAML file
When building the Settings value, we now also look for a 'settings.yaml'
file located in the data directory of pglift's code. Values from this
file may still be overridden through the SETTINGS environment variable.
The settings file is in YAML format, in order to allow comments or nicer
formatting of multi-line strings.