Move "install" command to main cli as "site-configure"
We add a "site-configure" command to the main CLI that performs what was
previously done by pglift.install's main section. The command is hidden,
as it is an internal command. We use the CLI context in do() and undo()
functions, thus getting the logger directly.

Thereafter, we rename the "install" module into "_install" to avoid
confusion (e.g. through command line completion) with the "instance"
module.