Improve parameters_from_model() composability
Previously, the decorator did not play well with other click
arguments/options defined with their own decorator on the callback
function.

In _decorators_from_model(), we now also yield the argument name in the
callback. Then, in parameters_from_model() we build the model (from
pydantic class) with only its respective parameters and finally invoke the
callback function with the arguments it expects. We add a doctest in
parameters_from_model() for future reference.