Run systemctl daemon-reload after change in unit files
After a manual change to systemd unit file. Following calls to systemctl
will trigger a warning "you must run systemctl daemon-reload".

I'm not sure this is required for all kind of changes in unit files but
at least this is required to rebuild dependencies (After/Before) and to
avoid the warning.

Also I noticed this may help to recover from interrupted functional
tests run (but there are still issues here).