Fix argspec_from_model() for nested optionals models
When building the ansible argspec from pydantic model, when a nested
model is Optional, required fields are marked as required and unset
fields are set to their "default" value.

This cause issues in case of an Optional nested model because required
field with no defaults fails the ansible validation.  Also a default
value on a field part of a model having another field required with no
default value will trigger model instantiation and fails the pydantic
validation.

Add a force_non_required flag in argspec_from_model() which is set to
True when nested model field is visited and is propagated to the whole
sub structure. Then when force_non_required is set, we don't set
"required" and "default" accordingly to the model, and rather rely on
pydantic validation.