Handle streaming replication slots
When creating a standby, we can now configure a slot name to be used.

When slot name is set, (re)create the replication slot either using
--create-slot options from pg_basebackup, or manually with
pg_create_physical_replication_slot() for postgresql 10 since
--create-slot is only available from postgresql >= 11.

Also note we are switching from --wal-method=fetch to the default
which is "stream". "stream" seems safer in all cases than "fetch", and
using slots require "stream" method.

Since we're adding a new parameter related to standby in InstanceSpec
and PostgreSQLInstance, and since there will probably be more of these
in the future, turn these parameters into a Standby class handling
system lookup.

Also I had to add PGPASSFILE to environ to be able to connect to the
primary when password is only available in passfile.