Fix test_list when multiple versions are available in tests
The test wasn't cleaning up the created directory, so it was failing on
mkdir() on second test.

Then we expect to found 1 instance in instance_list() call, but multiple
"test" instances will be found (one for each "pg_version" fixture
param). So just assert we found our instance, and basically check
"notAnInstanceDir" is not found.