Improve logging in CLI
We set the global log level to DEBUG so that all messages will be emitted
in the debug file handler.

Then we use the --log-level option to configure the level of the stream
handler, and set the default value to INFO.

Finally, we use distinct formatters for the debug file handler and for
the stream handler (the latter might be improved later on).