Reorganize and complete documentation
We now have two additional top-level sections in the documentation:

- "Setup and site configuration", describing the distribution and site
  administrator perspective, and
- "Operations", describing the operational perspective.

In the latter, we document all interfaces (command-line, Ansible and
Python) for each operational topic.

The "Setup and site configuration" section would still need more work,
especially to document settings management and configuration.

An API section is kept to document general aspects, not related to
specific operational context.

The objective is to have clearly defined entry-points in the
documentation for all kind of user profiles (distributor, site
administrator, operators, developers).