Make 'postgres_exporter {start,stop}' work with non-local instances
We only require the NAME of the postgres_exporter service which, for
local instance, would be the stanza <version>-<name>.

In tests, the postgres_exporter service is now bound to the existing
instance since we need to test the new service against a running
instance.

In systemd unit for postgres_exporter, we drop dependency on postgresql
service in order to allow starting a postgres_exporter service for a
non-local instance.