Display tasks title and status in CLI output
We introduce a "Displayer" object used in task.Runner and task.Task to
render tasks progress in the CLI. The Displayer is a context manager
which, upon enter, initialize the progress handler. The latter (which is
the same Displayer object) is then used upon each task run through its
handle() method to display task title.

The actual implementation in CLI uses rich's Live object and a grid to
stack tasks. Nested tasks are handled by using a 1-space indentation.

In order to limit mix of stderr (logging) and stdout (progress report),
we set the default log-level to WARNING.

In tests, we use no displayer.