Add support for running systemctl in --system mode
We add a systemd.user setting, on by default, which may be changed to
operate systemcl with --system option instead of --user.

We also add a sysuser setting, which default to the user running the
installation command. This is used to configure installed systemd unit
PostgreSQL with User/Group options in order to avoid running Postgres as
root (which is forbidden).