Honor pgpass setting when upgrading an instance
Upgrading an instance wasn't adding a pgpass entry even when pglift is
configured to do it, since surole_password was always None in new
instance manifest.

So configure the surole_password by using PGPASSWORD or by reading
passfile and looking for "old" instance password.

Also this allow us to set PGPASSWORD when calling pg_upgrade, so it will
not fail when entry for old instance is present in passfile and
PGPASSWORD is not supplied by the user.

Furthermore we don't need to temporary set trust authentication in new
instance anymore.