Allow to configure a "password_command" to set PGPASSWORD
When passwords aren't available in passfile and PGPASSWORD is not set
and pglift run postgresql commands in non interactive mode, we will
likely fail with "could not connect to server".

To handle such case, add a "password_command" under postgresql.auth
settings where we expect a path to a user-managed command returning
password suitable to use as PGPASSWORD environment variable.

Implement this in libpq_environ(), when PGPASSWORD is not present in
environment and password_command is not None, then we return the result
of the command as "PGPASSWORD".
Also handle this in pglift.db.superuser_connect().

In functional tests, this is quite the same scenario as "no_pgpass", to
avoid to add a new parametrization and define a password_command instead
of mocking os.environ["PGPASSWORD"].