Handle instances without postgres_exporter
We now allow the system Instance type to have a None 'prometheus' field,
thus properly accounting for the fact that an instance may not have the
service configured (instead of returning a wrong information in
system_lookup() previously).

In interface model, we also now allow None as 'prometheus' field, but
still defaults to a Prometheus object (with default port).

In unit tests, we build an Instance with and without a prometheus field
and adjust some tests accordingly.