Fix flapping test_backup_restore
The new version of pgbackrest now fail with "ERROR: [075]: unable to
find backup set with stop time less than <time>" when no backup
correspond to given time. Before it was just a warning, but not sure
what happened then...

https://pgbackrest.org/release.html#2.37

Sleep for 1 second in our test so we ensure pgbackrest will find our
backup.