Handle server reload and restart in instance.apply()
When a configuration parameter changed and a restart is needed for
changes to be effective, we return True in the (new) third item of the
return tuple and emit a warning log message.

When a reload is needed, we do it automatically, because it seems safe
and there is no service interrupt.

Special care is taken when the port has changed in order to avoid
querying the server for runtime settings as this would usually fail (at
least on local socket).