Do not try to start postgres_exporter if it is not enabled for an instance
We change enabled() into available() and now check for configuration
file existence in enabled(). Thus in instance_start() and
instance_stop(), we won't act if the postgres_exporter service is not
enabled for target instance.

This fixes a bug when trying to apply the following instance spec:

    name: noprometheus
    prometheus: null

In pgbackrest module, we also rename enabled() into available() for
consistency and to preserve component lookup from CLI code.