Separate display logic from transaction one
We drop the Runner class and instead define two context managers,
transaction() and displayer_installed(), each responsible for
transaction control and displayer configuration on the Task class.

Then in CLI, we only use the transaction control when it is safe, i.e.
only when we are initializing something, but not when we are updating
something (where rollback would destroy the previous state, which is not
desired).

In Ansible modules, we no longer use transaction control.