Allow to display and edit session default settings of a database
We add a "settings" field to Database model. When given a None value, no
change will be applied to an existing database object. Otherwise, items
from the dict will be use in an ALTER DATABASE query to either set a
configuration parameter, if a value is given or reset an existing
parameter if None is passed as a value.

This is not exposed from CLI (create or alter commands) because handling
of dict fields is not supported at the moment.