Build 'instance {create,alter,schema}' commands from registered plugins
We avoid using the global PLUGIN_MANAGER object, and thus
CompositeInstance, but now build the corresponding composite model from
the plugin manager currently in use (which might be different from the
global one, e.g. in tests).

This is achieved by defining a custom MultiCommand to register
'instance' sub-commands that cannot be defined statically but need this
composite instance model. This MultiCommand and the main 'instance'
group are merged thanks to click.CommandCollection.