Move 'postgres_exporter' and 'pgbackrest' commands in their plugin
In the main 'cli' module, we define a CLIGroup with get_command() and
list_commands() methods that loads commands or groups from plugins.

In this group, we need to retrieve the Obj value when it does not exists
(because the main cli() callback has not been invoked). In tests, this
will thus not use the context defined in tests (which in turns uses
custom settings).

As a consequence, we need to invoke plugin commands directly instead of
from the main 'cli' command.