Comment all entries of postgresql.conf at instance init
Some distribution packages ship a modified postgresql.conf, with
uncommented values. This might break our assumption that included files
on top of postgresql.conf will be accounted for and will create surprise
to the user.

So we get back to the upstream postgresql.conf file, which has all
values commented.

This resolves the inconsistency observed on redhat-systems, with PDGD
packages, where the 'log_directory' setting was not accounted for. And
similarly, the 'shared_buffers' setting was not accounted for on Debian.

In unit test about data-checksums, we now need to create the
postgresql.conf file in order the new code in instance.init() not to
crash.