Improve handling of invalid user settings
Instead of just logging (and thus possibly being ignored or getting
unnoticed by the user), we raise a proper exception and handle it in
CLI.

Though, if the settings file is empty, we just pass on.

Rework of commit 626467cf5c2f7c352c144b444e3feca5e0a3cacc.