Guess INSTANCE for 'instance' commands when there is only one
Similarly to -i/--instance for all other commands, we now guess the
INSTANCE argument if there is only one instance on system.

Accordingly, we document this behavior in commands.

In unit tests, when catching errors previously, we now run respective
tests with a parametrized argument to guess or give the instance
argument.