Write postgresql config changes after instance_configure() hooks run
Instead of writing site and user configuration files directly when the
instance gets configured, we store the configuration in memory and pass
it to instance_configure() hook along with the actual Instance object
(not the one that would take into account configuration changes). Then
after instance_configure() hook calls, we write configuration files.

We need a special case when the instance is being created and the
configuration is written directly in this case; hence the new
"_creating" argument to instance.configure().

The pgbackrest hooks needs adjustment to use the new configuration
passed from hook instead of querying it from instance. In constrast,
prometheus code seems okay already.