Always set password when available in "role alter"
Due to a condition in roles.alter() we were not updating the password
when role have already a password set.

Drop this condition and update password when a password is given on Role
object.

Modify test_apply() to check we can connect after password change (and
add missing login=True for this).

Add a test checking we can change the surole password.