Mark read-only interface model fields as such
We add the 'readOnly: true' JSON Schema annotation [1] to interface
model fields that cannot be modified, but only set at resource creation.
This is true for 'name' and 'version' of Instance resource, for
'name' of all resources and for password fields in Instance resource.

Accordingly, the 'exclude' option of parameters_from_model() decorator
is no longer useful but we now need to specify if an operation is a
creation or an update from CLI code.

On the other hand, since 'name' fields of Role and Database models are
now read-only, we need to add click.argument ourselves (but this is in
fact quite similar to instance commands implementation).

[1]: https://json-schema.org/draft/2020-12/json-schema-validation.html#rfc.section.9.4