Prompt for restart if needed upon 'pgconf edit'
We rework the whole command to call instances.configure() instead of
editing directly the "user.conf" file thus benefiting from the prompt to
restart the instance if needed.

In show_configuration_changes() we allow 'parameters' to be None as it
does not make much sense to provide a value from 'pgconf edit' case.