Restore missing shell completion of instance 'create' and 'alter' commands
We were returning None in InstanceCommands.get_command() when the click
context had no 'obj' bound because it's not possible to build a
composite instance model then. But as a consequence commands needed such
a model were not shown during shell completion. Instead, we use the base
Instance model when 'obj' is unset, thus getting a command name returned
from get_command() and fixing completion.