Define Instance's version field with an Enum
A new enum PostgreSQLVersion is introduced to list supported PostgreSQL
versions. It is based on StrEnum so that its value can be used as before
almost always when a string value was expected. This is used to define
the version field of interface and system model of Instance.

When defining choices for command-line option --version, we need to
explicitly build a list of PostgreSQLVersion because the enum is not a
sequence. CLI tests are extended for that.

In CLI, the neat benefit is that we get '--version [14|13|12|11|10]' in
help for create/alter instead of '--version VERSION' previously.
Similarly, the spec of "instance" Ansible module benefits from this.