Alert on creating instance if pgbackrest configuration exist
If there'is already a configuration file and maybe a backup
when creating an instance, we offer to abort creation or
overwrite files.
This can happen when the instance was deleted without deleting
pgbackrest's files (to keep backup) and we try to create one
with the same name.
If overwrite is chosen, we call pgbackrest revert_setup
to clear the existing configuration otherwise setup function will
use existing files without resetting them.