Use a sorted list type for privileges fields
The previous implementation using Set was not useful when exporting
models to JSON because the prettify() logic was not applied in that
case. On the other hand, it seems pretty complex to handle value sort
when exporting with .dict(). So we use a sorted list instead, along with
a validator to ensure sort.

In functional tests, we reorder privileges lists to make sure this has
no effect.