Implement replication lag computation for standby instances
We add instances.replication_lag() for this. It queries its primary for
its WAL LSN and then compute the lag from standby's replay value.
Connections to the primary and standby are done through the replication
user.

We test this in test_standby() by checking that the replication lag is
positive when the standby is out of sync and zero otherwise.