Let 'get' commands output either a table or some JSON
We add a table output to instance/role/database 'get' commands, and make
it the default. The YAML output is replaced by JSON, through the --json
option similarly to other commands.

Not that since there is only one row in tables output from 'get'
commands, we hide edges and lines (passing box=None to rich's Table).