Use CASCADE when creating or droping extensions
This automatically installs any extensions that the extension depends on
that are not already installed.

For example, `powa` extension depends on `btree_gist`. With CASCADE
`btree_gist` is automatically installed. Without, we get an
OperationalError.

Side effect, it will solve dependency error possibly caused by the order
of extensions passed to `instances.create_or_drop_extensions`.