Refactor functional tests 'settings' fixture
We introduce a 'postgresql_auth' parametrized fixture holding cases for
local (super-user) authentication: trust, password_command, pgpass.
Then, this is used by 'postgresql_settings' and 'settings' fixtures.
This is possible now that 'systemd' flag is controlled through a
command-line option of pytest, and no longer part of the 'settings'
fixture.