Make sure we keep extensions in order in instance._get
Previously we were sorting extensions after the concatenation of extensions from shared_preload_libraries and installed extensions.
The problem is that order in shared_preload_libraries is important. When altering an instance, the order may be changed if we do a sort.

The current commit makes sure that extensions read from shared_preload_libraries are kept in the same order