Add support for temboard-agent
At the moment, temboard agent must be on the same server as instance,
remote agent is currently unavailable.

This create and start a service for temboard-agent on the default port,
or the one specified through temboard-port option. Similarly, a password
for the temboard role used to connect to PostgreSQL may be specified.
The service template file is based on the template in temboard repository.

SSL files are self-signed and generated automatically. User can only
specify a directory to store them in settings.

You can specify temboard plugins in settings, by default it's set to
monitoring, dashboard and activity. Statements require pg_stat_statements
extension and pgconf and maintenance are redundant with pglift
features.

You must register the agent manually on the temboard UI.
The secret key is in the agent config file.

The temboard agent package installation has been added in CI.

Setup documentation has been added.