Make prometheus.execpath setting required
The previous default value was arbitrary and tied to the Debian package.
Since there is no obvious default value, let's require the field to be
set.

In tests, we turn the previous 'prometheus_available' fixture into
'prometheus_execpath' and let that detect well-known installations.
Accordingly, in RockyLinux CI image, we no longer rename the executable.