Catch invalid settings in CLI
Previously, the CLI would produce a traceback in case of invalid
settings because the pydantic.ValidationError is raised outside CLIGroup
(which handles validation errors for the interface model).

We now catch ValidationError when loading settings, on the Obj.
When going through shell completion, we catch InvalidSettingsError and
return no item in commands list; that's a compromise because displaying
an error message would actually produce a poorer user experience (as a
result of mixing stderr with readline).