Rework instance authentication configuration
We demote instances.instance_configure() from being a hook
implementation and make it a plain configure_auth() function called in
init(). Indeed, this was already a no-op when hba file was present so it
does not make sense to have implemented as a hook.

Now authentication configuration is done in apply(), when the instance
is being created; and replication role creation is performed there as
well, but outside of configure_auth() (previously instance_configure()),
thus explicitly.