Rework standby functional tests and setup
We define global standby_manifest and standby_instance fixtures that are
used in test_standby* at the moment but will be used elsewhere later on.

The previous test_standby() is split into two test cases. In
test_standby_instance(), we only check that the standby instance is
configured as expected. In test_standby_replication(), we check that
replication works. All this gets largely simplified, hopefully nothing
got lost.

We drop the 'slot' parameter and always use a slot (there is no
additional coverage in not using a slot).

Finally, the call of pg_drop_replication_slot() is dropped.