Allow to specify role password for postgres_exporter
We add a 'password' field to prometheus's Service and ServiceManifest
classes to store the password of postgres role. This role is also
renamed as 'prometheus' since 'monitoring' might be too generic and
conflict with other components.

This fixes the configuration of postgres_exporter on standby instances
with password authentication, hence the removal of xfail mark for
test_20_prometheus::test_standby.