Add DB & roles fields to Ansible module instance
We add the databases and roles field to the Ansible modules instance.
That way we can manage roles and databases directly from the instance
defintion.

For that we add the roles and databases fields to the instance models
and adapt the instance apply function to take care of creating the
corresponding roles and databases. This change also require to move the
definition of the database  and role class before the instance class
(they are now used in the instance class).

We change the tests for instance.apply (within test_10_instances.py)
with those new roles and databases fields.

The documentation is also edited to reference this new behavior.