Add support for PoWA
The current commit implements installation of PoWA extensions, and creation of dedicated role and database.
For now, it is meant to be used in a remote setup (data being collected remotely).

The role created for PoWA has superuser privileges in order to get
access to the query text.

When using CLI, the user is prompted for a password for the role if none
is already provided.