Set no target-action when restoring without a target
When calling pgbackrest.restore() with no target, we actually want to
restore from the latest backup *and* apply all available WALs. It seems
that pgbackrest needs to be called without any target-action in that
case.

In tests, we add a step checking this and complement the previous
restore() call with a check to make sure that restore reaches the
expected target (date).