Return a changed flag from instances.apply()
This flag indicates if something was done while applying requested
state, or if actual state already matches requested one. It is used in
the Ansible module to set the 'changed' flag.

We don't return configuration changes anymore, so their display in
Ansible module when diff mode is requested, is no longer available. This
diff mode was already partial and implemented in an awkward manner so
better not carry it along. If something is desired in this respect, we
should compute the diff of instance (interface) model before and after
call to apply(). Also note that changes to files would be detected if
we'd use AnsibleModule methods for file-system manipulations.

The system.Instance is no longer returned and callers are expected to
perform the lookup themselves.

In unit tests for the CLI, we drop return_value for mocked
instances.apply() since it's now wrong and wasn't used anyways.