Return a changed flag from roles.apply()
This flag indicates if something was done while applying requested
state, or if actual state already matches requested one. It is used in
the Ansible module to set the 'changed' flag.

As usual, we need special care when trying to connect to the database
with the super-user when its password is being modified because
dependent environment (typically, the password_command) might not have
been updated yet.