Configure new instance before running pg_upgrade command
This allows configuration to be already written on the new instance. This is important if extensions created in the source instance database require libraries to be loaded in shared_preload_libraries.

A new run_hooks parameter is added to instances.configure. It helps
preventing users to be created in hooks and avoid error like the
following during the pg_upgrade command:

Checking database user is the install user
Only the install user can be defined in the new cluster.