Check that a port is not used at instance creation
We define 'port' fields through the new Port type which has an
'available()' method to check that assigned value is not already in use
on host. This is then used in the new validate_ports() helper function
defined in models.interface module and called at instance creation.

This cannot be done through pydantic validation because there is no way
to distinguish between creation or update from there.

The Port type is mapped to an int, for Ansible model generation, though
this could be smarter by checking parent class in model converter code.

In transaction(), we now explicitly handle pydantic.ValidationError and
do not log them as it's typically a user error that should be handled
through the user interface.