Refactor nested model handling in _decorators_from_model()
We change the _prefix: str parameter into _parents: Tuple[str, ...] in
preparation for adding support for any nesting level.

We also move redefinition of 'argname' and 'modelname' later when there
are actually used.

We also simplify change from _ to - and vice-versa, hence adjustments
needed in metavar and when aliasing argument name in _cli_config.