Pass super-user password on standby creation in tests
Without this surole_password field in standby_manifest() fixture, and
dependent standby_instance() one, no pgpass entry will be added for the
super-user role. This is problematic because, when promoting the
instance, that password will not be available thus making it impossible
to connect. We extend test_standby_replication() to cover this case by
checking that we can connect to the instance after promotion and we add
test_standby_pgpass() to check that postgres role password is present
for standby port.

Finally, in test_99_drop.py::test_pgpass, we now use the
standby_instance fixture to make sure that we still get standby's
entry after dropping the primary (only for auth:pgpass flavor) and we
also check that this entry gets removed upon drop of the standby
instance (in standby_instance() fixture).