Rework pgpass checks in functional tests
Content comparison in test_99_drop.py::test_pgpass() was flaky because
it appears that passfile sort is not deterministic. So, in tests, we
remove the comment inserted in postgresql_settings() fixture and only
touch the file.

Accordingly, the password file will be dropped when the instance gets
dropped (by passfile.instance_drop() hook) so we now check this during
teardown of standby_instance() fixture.

Also, test_99_drop.py::test_pgpass() is now only run when authentication
mode is 'pgpass', thus removing inner conditions and making it simpler
to read.