Handle primary and stdby instances drop similarly in func tests
We add a standby_instance_dropped() fixture that is now used in
test_99_drop.py::test_pgpass(), which now checks that the password is
file dropped.

Now the primary and standby instances are handled similarly w.r.t. the
way they are dropped: either this happens when their respective
_dropped() fixture is called, or this happens at teardown of the main
fixture.