Always generate a "disable" flag for boolean parameters in CLI
This applies for instance to --login / --no-login in 'role create/alter'
commands.

The previous logic, to hide the disable flag when the parameter was
disabled by default, is removed and we now always produce a --no-<opt>
alternative to --<opt> in the CLI.

In unit tests, we drop 'address_shared' model field, which is now
redundant with 'address_primary'.