Automatically use "installed" test fixture
This is needed when running tests with --systemd. The fixture is
actually not useful as a test dependency, so we rename it with a leading
underscore.

Its code is adjusted to make sure it does what's expected when systemd
is requested; this comes with adjustments of _install.{do,undo}()'s
return value.

Prior to this commit, tests with --systemd were broken.